requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
docstring_parser==0.16
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
iniconfig==2.0.0
instructor==1.4.1
jiter==0.5.0
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
openai==1.43.1
packaging==24.1
pluggy==1.5.0
pydantic==2.9.0
pydantic_core==2.23.2
Pygments==2.18.0
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
pytest==8.3.2
python-dotenv==1.0.1
requests==2.32.3
rich==13.8.0
shellingham==1.5.4
slack_sdk==3.31.0
sniffio==1.3.1
tenacity==8.5.0
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
yarl==1.9.11
